#2452 Move KeyPropertyExpander and reduce its visibility
