+++ platform-dao/src/main/java/ua/com/fielden/platform/eql/stage1/queries/AbstractQuery1.java
@@ -28,8 +28,8 @@ import java.util.stream.Stream;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static java.util.Arrays.asList;
 import static ua.com.fielden.platform.eql.stage1.operands.Prop1.enhancePath;
-import static ua.com.fielden.platform.eql.stage2.KeyPropertyExpander.expand;
-import static ua.com.fielden.platform.eql.stage2.KeyPropertyExpander.isExpandable;
+import static ua.com.fielden.platform.eql.stage1.queries.KeyPropertyExpander.expand;
+import static ua.com.fielden.platform.eql.stage1.queries.KeyPropertyExpander.isExpandable;
 import static ua.com.fielden.platform.eql.stage2.conditions.Conditions2.EMPTY_CONDITIONS;
 import static ua.com.fielden.platform.eql.stage2.conditions.Conditions2.conditions;
 import static ua.com.fielden.platform.eql.stage2.sundries.GroupBys2.EMPTY_GROUP_BYS;
